some refactor, init testify, shortener test
