+++ cmd/shortener/shortener_test.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"github.com/stretchr/testify/require"
+	"testing"
+)
+
+func TestShorting(t *testing.T) {
+	require.NotEmpty(t, Shorting())
+	require.Len(t, Shorting(), 8)
+}
